Add files via upload
recomb.c is a c program that estimates the longest straight line with a slope <0.025 cM/Mb that has no points 2 cM from the line. recomb.tsv is the input dataset for the recombination genetic celery map as a function of the physical map of celery cv. Challenger. recomb.out is the output.